refactor(async): name batch control module precisely
Keep trainer-owned batch admission and publication control distinct from the RL policy implementation.
